Refactoring to make output additions more clearly navigable
